+++ gammapy/makers/tests/test_map.py
@@ -3,12 +3,13 @@ import pytest
 from numpy.testing import assert_allclose
 import numpy as np
 import astropy.units as u
+from regions import PointSkyRegion
 from astropy.coordinates import SkyCoord
 from gammapy.data import DataStore
 from gammapy.datasets import MapDataset
 from gammapy.irf import EDispMap, EDispKernelMap
 from gammapy.makers import MapDatasetMaker, SafeMaskMaker
-from gammapy.maps import Map, MapAxis, WcsGeom
+from gammapy.maps import Map, MapAxis, WcsGeom, RegionGeom
 from gammapy.utils.testing import requires_data
 
 
@@ -204,3 +205,116 @@ def test_make_meta_table(observations):
     assert_allclose(map_dataset_meta_table["RA_PNT"], 267.68121338)
     assert_allclose(map_dataset_meta_table["DEC_PNT"], -29.6075)
     assert_allclose(map_dataset_meta_table["OBS_ID"], 110380)
+
+
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+import pytest
+import numpy as np
+from numpy.testing import assert_allclose
+from astropy.coordinates import Angle, SkyCoord
+from gammapy.data import DataStore
+from gammapy.maps import MapAxis
+from gammapy.utils.testing import requires_data
+
+
+@pytest.fixture(scope="session")
+def data_store():
+    return DataStore.from_dir("$GAMMAPY_DATA/hess-dl3-dr1/")
+
+
+@requires_data()
+@pytest.mark.parametrize(
+    "pars",
+    [
+        {
+            "energy": None,
+            "rad": None,
+            "energy_shape": (32,),
+            "psf_energy": 865.9643,
+            "rad_shape": (144,),
+            "psf_rad": 0.0015362848,
+            "psf_exposure": 3.14711e12,
+            "psf_value_shape": (32, 144),
+            "psf_value": 4369.96391,
+        },
+        {
+            "energy": MapAxis.from_energy_bounds(1, 10, 100, "TeV", name="energy_true"),
+            "rad": None,
+            "energy_shape": (101,),
+            "psf_energy": 1412.537545,
+            "rad_shape": (144,),
+            "psf_rad": 0.0015362848,
+            "psf_exposure": 4.688142e12,
+            "psf_value_shape": (101, 144),
+            "psf_value": 3726.58798,
+        },
+        {
+            "energy": None,
+            "rad": MapAxis.from_nodes(np.arange(0, 2, 0.002) * u.deg, name="theta"),
+            "energy_shape": (32,),
+            "psf_energy": 865.9643,
+            "rad_shape": (1000,),
+            "psf_rad": 0.000524,
+            "psf_exposure": 3.14711e12,
+            "psf_value_shape": (32, 1000),
+            "psf_value": 25888.5047,
+        },
+        {
+            "energy": MapAxis.from_energy_bounds(1, 10, 100, "TeV", name="energy_true"),
+            "rad": MapAxis.from_nodes(np.arange(0, 2, 0.002) * u.deg, name="theta"),
+            "energy_shape": (101,),
+            "psf_energy": 1412.537545,
+            "rad_shape": (1000,),
+            "psf_rad": 0.000524,
+            "psf_exposure": 4.688142e12,
+            "psf_value_shape": (101, 1000),
+            "psf_value": 22723.879272,
+        },
+    ],
+)
+def test_make_psf(pars, data_store):
+    energy_axis = pars["energy"]
+    rad_axis = pars["rad"]
+
+    if energy_axis is None:
+        energy_axis = MapAxis.from_energy_bounds(1, 10, 100, unit="TeV", name="energy_true")
+
+    if rad_axis is None:
+        rad_axis = MapAxis.from_bounds(0, 10, 100, unit="deg", name="theta")
+
+    geom = RegionGeom(
+        region=PointSkyRegion(SkyCoord(83.63, 22.01, unit="deg")),
+        axes=[rad_axis, energy_axis]
+    )
+
+    maker = MapDatasetMaker()
+
+    psf_map = maker.make_psf(geom, data_store.obs(23523))
+    psf = psf_map.get_energy_dependent_table_psf()
+
+    assert psf.energy.unit == "GeV"
+    assert psf.energy.shape == pars["energy_shape"]
+    assert_allclose(psf.energy.value[15], pars["psf_energy"], rtol=1e-3)
+
+    assert psf.rad.unit == "rad"
+    assert psf.rad.shape == pars["rad_shape"]
+    assert_allclose(psf.rad.value[15], pars["psf_rad"], rtol=1e-3)
+
+    assert psf.exposure.unit == "cm2 s"
+    assert psf.exposure.shape == pars["energy_shape"]
+    assert_allclose(psf.exposure.value[15], pars["psf_exposure"], rtol=1e-3)
+
+    assert psf.psf_value.unit == "sr-1"
+    assert psf.psf_value.shape == pars["psf_value_shape"]
+    assert_allclose(psf.psf_value.value[15, 50], pars["psf_value"], rtol=1e-3)
+
+
+@requires_data()
+def test_make_mean_psf(data_store):
+    position = SkyCoord(83.63, 22.01, unit="deg")
+
+    observations = data_store.get_observations([23523, 23526])
+    psf = make_mean_psf(observations, position=position)
+
+    assert not np.isnan(psf.psf_value.value).any()
+    assert_allclose(psf.psf_value.value[22, 22], 12206.1665)
